Added form stories
